start profile creation and writing
